requirements.txt
google-adk>=1.31.0
google-genai>=1.73.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
yfinance>=0.2.0
streamlit>=1.40.0